update param type in scanhash_sha512

+++ miner.h
@@ -345,7 +345,7 @@ bool stratum_update_dataset(struct stratum_ctx *sctx, const char *user, const ch
 uint64_t* updateLookupTBL(uint8_t seedhash[OFF_CYCLE_LEN+SKIP_CYCLE_LEN][32],uint64_t *plookupTbl,int plen);
 void truehashTableInit(uint64_t *tableLookup,int tlen);
 bool dataset_hash(uint8_t hash[64], uint64_t *data,int len);
-int scanhash_sha512(int thr_id, const uint64_t *dataset,int dlen,uint8_t hash[HEADSIZE], uint32_t target[TARGETLEN],
+int scanhash_sha512(int thr_id, const uint64_t *dataset,int dlen,uint8_t hash[HEADSIZE], uint8_t target[TARGETLEN],
 				uint64_t *nonce,uint64_t max_nonce, uint64_t *hashes_done);
 
 struct thread_q;
